moved coverages in doc folder
